+++ src/fallback.ts
@@ -28,17 +28,18 @@ import {
   GoogleAuthOptions,
   BaseExternalAccountClient,
 } from 'google-auth-library';
-import * as objectHash from 'object-hash';
 import {OperationsClientBuilder} from './operationsClient';
 import {GrpcClientOptions, ClientStubOptions} from './grpc';
 import {GaxCall, GRPCCall} from './apitypes';
-import {Descriptor} from './descriptor';
+import {Descriptor, StreamDescriptor} from './descriptor';
 import {createApiCall as _createApiCall} from './createApiCall';
 import {FallbackServiceError} from './googleError';
 import * as fallbackProto from './fallbackProto';
 import * as fallbackRest from './fallbackRest';
 import {isNodeJS} from './featureDetection';
 import {generateServiceStub} from './fallbackServiceStub';
+import {StreamType} from './streamingCalls/streaming';
+import * as objectHash from 'object-hash';
 
 export {FallbackServiceError};
 export {PathTemplate} from './pathTemplate';
@@ -134,7 +135,7 @@ export class GrpcClient {
   }
 
   loadProtoJSON(json: protobuf.INamespace, ignoreCache = false) {
-    const hash = objectHash(json);
+    const hash = objectHash(JSON.stringify(json)).toString();
     const cached = GrpcClient.protoCache.get(hash);
     if (cached && !ignoreCache) {
       return cached;
@@ -360,10 +361,14 @@ export function createApiCall(
   settings: gax.CallSettings,
   descriptor?: Descriptor
 ): GaxCall {
-  if (descriptor && 'streaming' in descriptor) {
+  if (
+    descriptor &&
+    'streaming' in descriptor &&
+    (descriptor as StreamDescriptor).type !== StreamType.SERVER_STREAMING
+  ) {
     return () => {
       throw new Error(
-        'The gRPC-fallback client library (e.g. browser version of the library) currently does not support streaming calls.'
+        'The gRPC-fallback client library (e.g. browser version of the library) currently does not support client-streaming or bidi-stream calls.'
       );
     };
   }