+++ test/system-test/showcase-server.ts
@@ -31,7 +31,7 @@ export class ShowcaseServer {
     const testDir = path.join(process.cwd(), '.showcase-server-dir');
     const platform = process.platform;
     const arch = process.arch === 'x64' ? 'amd64' : process.arch;
-    const showcaseVersion = process.env['SHOWCASE_VERSION'] || '0.6.1';
+    const showcaseVersion = process.env['SHOWCASE_VERSION'] || '0.17.0';
     const tarballFilename = `gapic-showcase-${showcaseVersion}-${platform}-${arch}.tar.gz`;
     const fallbackServerUrl = `https://github.com/googleapis/gapic-showcase/releases/download/v${showcaseVersion}/${tarballFilename}`;
     const binaryName = './gapic-showcase';