+++ test/unit/grpc-fallback.ts
@@ -27,6 +27,7 @@ import * as protobuf from 'protobufjs';
 import * as sinon from 'sinon';
 import {echoProtoJson} from '../fixtures/echoProtoJson';
 import {GrpcClient} from '../../src/fallback';
+import {GoogleError} from '../../src';
 
 // @ts-ignore
 const hasAbortController = typeof AbortController !== 'undefined';
@@ -127,9 +128,10 @@ describe('createStub', () => {
     assert.strictEqual(typeof echoStub.echo, 'function');
     assert.strictEqual(typeof echoStub.pagedExpand, 'function');
     assert.strictEqual(typeof echoStub.wait, 'function');
+    assert.strictEqual(typeof echoStub.close, 'function');
 
-    // There should be 6 methods for the echo service
-    assert.strictEqual(Object.keys(echoStub).length, 6);
+    // There should be 7 methods for the echo service + 1 close method.
+    assert.strictEqual(Object.keys(echoStub).length, 8);
 
     // Each of the service methods should take 4 arguments (so that it works with createApiCall)
     assert.strictEqual(echoStub.echo.length, 4);
@@ -142,9 +144,10 @@ describe('createStub', () => {
     assert.strictEqual(typeof echoStub.echo, 'function');
     assert.strictEqual(typeof echoStub.collect, 'function');
     assert.strictEqual(typeof echoStub.chat, 'function');
+    assert.strictEqual(typeof echoStub.close, 'function');
 
-    // There should be 6 methods for the echo service
-    assert.strictEqual(Object.keys(echoStub).length, 6);
+    // There should be 7 methods for the echo service + 1 close method.
+    assert.strictEqual(Object.keys(echoStub).length, 8);
 
     // Each of the service methods should take 4 arguments (so that it works with createApiCall)
     assert.strictEqual(echoStub.echo.length, 4);
@@ -310,12 +313,82 @@ describe('grpc-fallback', () => {
         },
       })
     );
-
     gaxGrpc.createStub(echoService, stubOptions).then(echoStub => {
       echoStub.echo(requestObject, {}, {}, (err?: Error) => {
-        assert(err instanceof Error);
+        assert(err instanceof GoogleError);
         assert.strictEqual(err.message, expectedMessage);
-        assert.strictEqual(JSON.stringify(err), JSON.stringify(expectedError));
+        assert.strictEqual(err.code, expectedError.code);
+        assert.strictEqual(
+          JSON.stringify(err.statusDetails),
+          JSON.stringify(expectedError.details)
+        );
+        done();
+      });
+    });
+  });
+
+  it('should promote ErrorInfo if exist in fallback-rest error', done => {
+    const requestObject = {content: 'test-content'};
+    // example of an actual google.rpc.Status error message returned by Translate API
+    const errorInfo = {
+      '@type': 'type.googleapis.com/google.rpc.ErrorInfo',
+      reason: 'SERVICE_DISABLED',
+      domain: 'googleapis.com',
+      metadata: {
+        service: 'translate.googleapis.com',
+        consumer: 'projects/123',
+      },
+    };
+    const serverError = {
+      error: {
+        code: 403,
+        message:
+          'Cloud Translation API has not been used in project 123 before or it is disabled. Enable it by visiting https://console.developers.google.com/apis/api/translate.googleapis.com/overview?project=455411330361 then retry. If you enabled this API recently, wait a few minutes for the action to propagate to our systems and retry.',
+        status: 'PERMISSION_DENIED',
+        details: [
+          {
+            '@type': 'type.googleapis.com/google.rpc.Help',
+            links: [
+              {
+                description: 'Google developers console API activation',
+                url: 'https://console.developers.google.com/apis/api/translate.googleapis.com/overview?project=455411330361',
+              },
+            ],
+          },
+          errorInfo,
+        ],
+      },
+    };
+    const opts = {
+      auth: authStub,
+      fallback: 'rest',
+    };
+    // @ts-ignore incomplete options
+    gaxGrpc = new GrpcClient(opts);
+    //@ts-ignore
+    sinon.stub(nodeFetch, 'Promise').returns(
+      Promise.resolve({
+        ok: false,
+        arrayBuffer: () => {
+          return Promise.resolve(Buffer.from(JSON.stringify(serverError)));
+        },
+      })
+    );
+    gaxGrpc.createStub(echoService, stubOptions).then(echoStub => {
+      echoStub.echo(requestObject, {}, {}, (err?: Error) => {
+        assert(err instanceof GoogleError);
+        assert.strictEqual(
+          JSON.stringify(err.statusDetails),
+          JSON.stringify(serverError['error']['details'])
+        );
+        assert.strictEqual(err.code, 7);
+        assert.strictEqual(err.message, serverError['error']['message']);
+        assert.strictEqual(err.reason, errorInfo.reason);
+        assert.strictEqual(err.domain, errorInfo.domain);
+        assert.strictEqual(
+          JSON.stringify(err.errorInfoMetadata),
+          JSON.stringify(errorInfo.metadata)
+        );
         done();
       });
     });
@@ -333,5 +406,14 @@ describe('grpc-fallback', () => {
 
     // @ts-ignore
     assert.strictEqual(createdAbortControllers[0].abortCalled, true);
+  });
+
+  it('should have close method', done => {
+    // @ts-ignore
+    sinon.stub(nodeFetch, 'Promise').returns(Promise.resolve({}));
+    gaxGrpc.createStub(echoService, stubOptions).then(stub => {
+      stub.close({}, {}, {}, () => {});
+      done();
+    });
   });
 });